refactor: dashboard-header를 header/index.tsx로 이동
